fix: 자잘한 디자인 수정및 코드 리팩토링
